Reinstate double deprecation test logic
This was lost in a previous change, but in the interest of not holding
up the merge of the functional part of it I'm fixing the test in this
followup. The new version of the test verifies that the rule isn't
overwritten by a second call to load_rules, which is what would
happen if we called the deprecation handler again.

Change-Id: Idc9dd353463da63a2108782f7b141d1d8014d180