Don't reset rules without overwriting
If an user uses Enforcer without overwriting (Enforcer(overwrite=False))
we  should not reset rules and only update loaded rules.
Enforcer without overwriting is a weird behavior, but it is supported at this moment.
Maybe it will be eliminated in future because it's misleading.
Operator cannot conclude what rules are loaded by simply looking in config files.

Change-Id: I2871407f8c7417a016415ccc166c1f37a9e17908
Closes-Bug: 1943584