Expand set_defaults() to set other config default value
Currently set_defaults() is only able to set the default
value of policy_file config option. In future, for example
scope config option like enforce_scope also needs to be override
the default value per service (service ready with scope enable
can set it to True and for other services it will be False as default
in oslo.policy).

To allow override the other config option, let's expand the existing
set_defaults() method to do so.

Change-Id: I72120efb7c55aab82b765237904c9ae6e91f6b6f