Drop python 3.8 support
This drop the python 3.8 support. Also, remove
the python_requires flag from setup.cfg which
stop installing it on python version lower
than the min supported. Let's allow to install
and as long as it work for older python it is
ok. User can get to know the supported python
versions from classifier also.

Change-Id: Ic77001276e09288debc27258728dc039193bf6cb