parser: Remove use of metaclass
This code is too clever by half, especially given we do not expect
or support registered additional reducers. Simplify things greatly
by doing things statically.

Tests are updated to reflect both the removal of the metaclass and the
need to use the inbuilt "reducers" (and therefore the expected token
types).

Change-Id: I4d37f6c12aef4c8423db12118480637c8fe1f83f
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>